Mecanum/requirements.txt
websockets
python-dotenv
# smbus